\name{data.ISOPAL}
\docType{data}

\alias{data.ISOPAL}

\title{Dataset: list of rodent association}
\description{
   data.ISOPAL is a list of data used for the package ISOPAL.
}
\usage{data(data.ISOPAL)}
\format{
    \itemize{
        \item\code{data.ISOPAL} is a matrix. 
    }
    
}
\references{
    \itemize{

		}
        }
\keyword{datasets}

\examples{
data(data.ISOPAL)
}

