\name{ISOPAL}
\alias{ISOPAL}

\title{ISOPAL}

\description{This package is a R shiny application dedicated to the generate interpolation maps based on rodent associations, bioclim methods and GMML equations.}

\usage{
ISOPAL()
}

\details{This is a launcher function for the `ISOPAL` application.}

\value{Launch the `ISOPAL` Shiny application.}

\author{Aurelien Royer <aurelien.royer at u-bourgogne.fr>}

\seealso{
	\code{shiny}
	\code{plotly}
	\code{IsoriX}
}

\references{
Royer, A., DOI: 10.5281/zenodo.xxxx.
}

\examples{
	if(interactive()){ISOPAL()}
}
